Inline mozilla definitions
